[Pages] All Pages Templates and start of Wix Client calls
