+++ .env.development
@@ -1,2 +1,5 @@
 # this is a template, copy to .env.local with proper values for local development (do not store API Key in git)
 NEXT_PUBLIC_WIX_CLIENT_ID=3bca9b0d-398e-4c47-b2e2-263d4ccd7458
+WIX_CLIENT_ID=your-client-id
+WIX_CLIENT_SECRET=your-client-secret
+WIX_API_KEY=your-api-key
