--- app/page-components/EventPageComponent/EventPageComponent.module.css
@@ -1,89 +1,0 @@
-.postContainer {
-  margin: 0 auto;
-  max-width: 960px;
-  display: flex;
-  flex-direction: column;
-  margin-top: 1em;
-
-  .preHeader {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-top: 1em;
-  }
-
-  .postHeader {
-    width: 100%;
-    display: flex;
-    margin-top: 2em;
-
-    .imageAndLinks {
-      flex-shrink: 0;
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      margin-right: 2em;
-      width: 280px;
-    }
-    .socialIcons {
-      width: 90%;
-      display: flex;
-      margin-top: 0.5em;
-      justify-content: center;
-
-      svg {
-        margin-left: 10px;
-      }
-    }
-    .detailsColumn {
-      flex-grow: 1;
-      width: auto;
-      display: flex;
-      flex-direction: column;
-    }
-  }
-
-  .recommandations {
-    border-top: 1px solid #d9d9d9;
-    display: flex;
-    /* justify-content: center; */
-    align-items: center;
-    margin-top: 0.5em;
-    padding-top: 0.5em;
-    width: 100%;
-  }
-
-  .foresight,
-  .domains,
-  .projects,
-  .posts,
-  .projectResults,
-  .events,
-  .affiliations {
-    display: flex;
-    align-items: center;
-    margin-top: 1em;
-    flex-wrap: wrap;
-    margin-bottom: 1em;
-  }
-
-  .foresightTitle,
-  .domainTitle,
-  .projectsTitle,
-  .affiliationsTitle {
-    font-weight: bold;
-    border-bottom: 1px solid #d9d9d9;
-    padding-bottom: 12px;
-    margin-bottom: 1em;
-  }
-
-  .postContent {
-    margin: 0 auto;
-    display: flex;
-    flex-direction: row;
-    /* align-items: center; */
-  }
-  .domains {
-    margin-bottom: 2em;
-  }
-}
