+++ app/page-components/OrganisationPageComponent/components/AffiliationsComponent/AffiliationsComponent.module.css
@@ -1,0 +1,18 @@
+.personWorkplace {
+  display: flex;
+  flex-direction: column;
+  margin-top: 1em;
+  width: 100%;
+
+  .personWorkplaceItem {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+  }
+}
+.affiliationsTitle {
+  font-weight: bold;
+  border-bottom: 1px solid #d9d9d9;
+  padding-bottom: 12px;
+  margin-bottom: 1em;
+}
