+++ app/page-components/OrganisationPageComponent/components/HeaderComponent/HeaderComponent.tsx
@@ -1,0 +1,131 @@
+import Image from 'next/image';
+import style from './HeaderComponent.module.css';
+import classNames from 'classnames';
+import Typography from '@app/shared-components/Typography/Typography';
+import Tag, { TagProps } from '@app/shared-components/Tag/Tag';
+
+export type HeaderComponentProps = {
+  organisation: {
+    title: string;
+    tagLine: string;
+    image: string;
+    personPopularity: number;
+    pageType: 'Post' | 'Event' | 'Project Result';
+    description: string;
+    countryTag: {
+      tagText: string;
+      tagCounter: number;
+    };
+    recommendations: {
+      number: number;
+      images: string[];
+    };
+    views: number;
+    domains: Array<TagProps>;
+  };
+};
+
+const HeaderComponent: React.FC<HeaderComponentProps> = ({ organisation }) => {
+  return (
+    <div className={classNames(style.personHeader)}>
+      <div className={style.imageAndSocialColumn}>
+        <Image
+          src={organisation.image}
+          width={147}
+          height={147}
+          className={classNames('rounded-full')}
+          alt={`User Avatar - ${organisation.title}`}
+        />
+        {/* Social Icons */}
+        {/* Social Icons */}
+        <div className={style.socialIcons}>
+          {/* Linkedin */}
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            className=" rounded-full"
+            fill="currentColor"
+            width={32}
+            height={32}
+            style={{
+              color: '#fff',
+              backgroundColor: '#0077b5',
+              padding: '5px',
+            }}
+            viewBox="0 -7 23 40"
+          >
+            <path d="M4.98 3.5c0 1.381-1.11 2.5-2.48 2.5s-2.48-1.119-2.48-2.5c0-1.38 1.11-2.5 2.48-2.5s2.48 1.12 2.48 2.5zm.02 4.5h-5v16h5v-16zm7.982 0h-4.968v16h4.969v-8.399c0-4.67 6.029-5.052 6.029 0v8.399h4.988v-10.131c0-7.88-8.922-7.593-11.018-3.714v-2.155z" />
+          </svg>
+          {/* Link */}
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            fill="none"
+            viewBox="0 0 24 24"
+            strokeWidth={1.5}
+            stroke="currentColor"
+            className="w-6 h-6 rounded-full p-1"
+            style={{ backgroundColor: '#9d8fdf', color: 'white' }}
+          >
+            <path
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              d="M13.19 8.688a4.5 4.5 0 011.242 7.244l-4.5 4.5a4.5 4.5 0 01-6.364-6.364l1.757-1.757m13.35-.622l1.757-1.757a4.5 4.5 0 00-6.364-6.364l-4.5 4.5a4.5 4.5 0 001.242 7.244"
+            />
+          </svg>
+        </div>
+        {/* Views */}
+        <Typography
+          data-after="2153"
+          tag="p"
+          className="text-sm text-gray-800 my-3 after:content-[attr(data-after)]]"
+        >
+          {organisation.views} views
+        </Typography>
+      </div>
+      <div className={style.detailsColumn}>
+        {/* Organisation Info Name */}
+        <Typography tag="h1" className=" text-gray-800">
+          {organisation.title.replace(/_/g, ' ')}
+          {/* Organisation Popularity */}
+          <span
+            data-after={organisation.personPopularity}
+            className="after:content-[attr(data-after)] text-lg relative top-[-30px] ml-1 text-gray-500 dark:text-gray-400"
+          ></span>
+        </Typography>
+        {/* Founded */}
+        <div className="flex items-center my-2">
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            fill="none"
+            viewBox="0 0 24 24"
+            strokeWidth={1.5}
+            stroke="currentColor"
+            className="w-6 h-6"
+          >
+            <path
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              d="M12 21v-8.25M15.75 21v-8.25M8.25 21v-8.25M3 9l9-6 9 6m-1.5 12V10.332A48.36 48.36 0 0012 9.75c-2.551 0-5.056.2-7.5.582V21M3 21h18M12 6.75h.008v.008H12V6.75z"
+            />
+          </svg>
+          <Typography tag="span" className="text-gray-800 italic text-xs ml-2">
+            Founded in 1479
+          </Typography>
+        </div>
+        {/* Tagline */}
+        <Typography tag="h3" className="text-gray-800 italic">
+          {organisation.tagLine}
+        </Typography>
+        {/* Organisation domains */}
+        <div className={style.domains}>
+          {organisation.domains.slice(0, 3).map((domain) => (
+            <Tag key={domain.tagText} {...domain} />
+          ))}
+        </div>
+        {/* Organisation Country */}
+        <Tag {...organisation.countryTag} />
+      </div>
+    </div>
+  );
+};
+
+export default HeaderComponent;
