+++ app/page-components/PersonPageComponent/components/AffiliationsComponent/AffiliationsComponent.tsx
