--- app/page-components/PostPageComponent/components/TagListComponent/TagListComponent.module.css
@@ -1,13 +1,0 @@
-.tagListContainer {
-  display: flex;
-  align-items: center;
-  margin-top: 1em;
-  flex-wrap: wrap;
-  margin-bottom: 1em;
-}
-.tagListTitle {
-  font-weight: bold;
-  border-bottom: 1px solid #d9d9d9;
-  padding-bottom: 12px;
-  margin-bottom: 1em;
-}
