+++ app/page-components/ProjectPageComponent/components/AffiliationsComponent/AffiliationsComponent.tsx
@@ -1,0 +1,57 @@
+import Image from 'next/image';
+import style from './AffiliationsComponent.module.css';
+import classNames from 'classnames';
+import Tag, { TagProps } from '@app/shared-components/Tag/Tag';
+import Typography from '@app/shared-components/Typography/Typography';
+
+export type AffiliationsComponentProps = {
+  afiliations: Array<
+    TagProps & {
+      arole: string;
+    }
+  >;
+  title?: string;
+};
+
+const AffiliationsComponent: React.FC<AffiliationsComponentProps> = ({
+  afiliations,
+  title,
+}) => {
+  return (
+    <section className={style.personWorkplace}>
+      <Typography
+        tag="h2"
+        className={classNames(
+          'text-gray-800 w-full my-4',
+          style.affiliationsTitle
+        )}
+      >
+        {title}
+      </Typography>
+
+      {afiliations.map((affilitiation) => (
+        <div
+          key={`affiliation-${affilitiation.tagText}`}
+          className={classNames(style.personWorkplaceItem)}
+        >
+          {affilitiation.arole && (
+            <Typography
+              tag="span"
+              className={classNames('text-gray-400', 'mr-2')}
+            >
+              {affilitiation.arole}:
+            </Typography>
+          )}
+          <Tag
+            tagText={affilitiation.tagText}
+            tagCounter={affilitiation.tagCounter}
+            href={affilitiation.href}
+            thumbnail={affilitiation.thumbnail}
+          />
+        </div>
+      ))}
+    </section>
+  );
+};
+
+export default AffiliationsComponent;
