--- app/page-components/ProjectResultPageComponent/ProjectResultPageComponent.module.css
@@ -1,131 +1,0 @@
-.personContainer {
-  margin: 0 auto;
-  max-width: 960px;
-  display: flex;
-  flex-direction: column;
-  margin-top: 1em;
-
-  .preHeader {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-top: 1em;
-  }
-
-  .personHeader {
-    display: flex;
-    margin-top: 2em;
-
-    .imageAndButtons {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      margin-right: 2em;
-      flex-shrink: 0;
-
-      .projectResultImage {
-        outline: lightGray solid 1px;
-        border-radius: 8px;
-        transition: all 0.3s ease-in-out;
-      }
-      .projectResultImage:hover {
-        cursor: pointer;
-        box-shadow: rgb(75, 145, 243) 0px 0px 5px;
-      }
-    }
-    .downloadAndViews {
-      width: 100%;
-      display: flex;
-      margin-top: 0.5em;
-      justify-content: center;
-    }
-    .detailsColumn {
-      display: flex;
-      flex-direction: column;
-    }
-
-    .recommandations {
-      border-top: 1px solid #d9d9d9;
-      display: flex;
-      /* justify-content: center; */
-      align-items: center;
-      margin-top: 0.5em;
-      padding-top: 0.5em;
-      width: 100%;
-    }
-  }
-
-  .personWorkplace {
-    display: flex;
-    flex-direction: column;
-    margin-top: 1em;
-
-    .personWorkplaceItem {
-      position: relative;
-      margin: 0.5em 0;
-    }
-  }
-
-  .recommandations {
-    display: flex;
-    /* justify-content: center; */
-    align-items: center;
-    margin-top: 0.5em;
-  }
-
-  .foresight,
-  .domains,
-  .coordinators,
-  .posts,
-  .projectResults,
-  .events,
-  .affiliations,
-  .team,
-  .organisations {
-    display: flex;
-    align-items: center;
-    margin-top: 1em;
-    flex-wrap: wrap;
-    margin-bottom: 1em;
-  }
-
-  .coordinatorsContent,
-  .teamContent,
-  .organisationsContent {
-    position: relative;
-  }
-
-  .organisationsContent {
-    display: flex;
-    flex-direction: column;
-    margin-top: 1em;
-    margin-bottom: 1em;
-  }
-  .organisationContentItem {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-  }
-
-  .foresightTitle,
-  .domainTitle,
-  .projectsTitle,
-  .affiliationsTitle {
-    font-weight: bold;
-    border-bottom: 1px solid #d9d9d9;
-    padding-bottom: 12px;
-    margin-bottom: 1em;
-  }
-
-  .affiliationsContent {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    margin-top: 1em;
-  }
-  .affiliationsItem {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-  }
-}
