+++ app/page-components/shared-page-components/ExternalLinksComponent/ExternalLinksComponent.tsx
