+++ app/page-components/shared-page-components/MiniPagesListComponent/MiniPagesListComponent.module.css
@@ -1,0 +1,9 @@
+.foresightTitle,
+.domainTitle,
+.projectsTitle,
+.affiliationsTitle {
+  font-weight: bold;
+  border-bottom: 1px solid #d9d9d9;
+  padding-bottom: 12px;
+  margin-bottom: 1em;
+}
