+++ app/project/[slug]/page.tsx
@@ -7,7 +7,7 @@ export default function PostPage({ params }: any) {
   console.log('Post Page Params', params.slug);
   return (
     <div className={classNames('w-full')}>
-      <ProjectPageComponent />
+      <ProjectPageComponent pageTitle={params.slug} />
     </div>
   );
 }
