+++ package.json
@@ -26,7 +26,8 @@
     "react-dom": "^18.2.0",
     "react-draft-wysiwyg": "^1.15.0",
     "react-scripts": "^5.0.1",
-    "react-select": "^5.8.0"
+    "react-select": "^5.8.0",
+    "wix-sdk": "^1.90.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.24.1",
